requirements.txt
pyjwt
pwdlib
sqlalchemy
psycopg[binary]
fastapi
uvicorn
python-dotenv
python-multipart